+++ cli/subcommands/src/cargo_hax.rs
@@ -169,9 +169,12 @@ impl HaxMessage {
             } => {
                 let mut _rctx = None;
                 let rctx = rctx.unwrap_or_else(|| _rctx.get_or_insert(ReportCtx::default()));
-                diagnostic.with_message(rctx, &working_dir, Level::Error, |msg| {
-                    eprintln!("{}", renderer.render(msg))
-                });
+                diagnostic.with_message(
+                    rctx,
+                    working_dir.as_ref().map(PathBuf::as_path),
+                    Level::Error,
+                    |msg| eprintln!("{}", renderer.render(msg)),
+                );
             }
             Self::EngineNotFound {
                 is_opam_setup_correctly,
@@ -258,8 +261,8 @@ impl HaxMessage {
 fn run_engine(
     haxmeta: HaxMeta<hax_frontend_exporter::ThirBody>,
     id_table: id_table::Table,
-    working_dir: PathBuf,
-    manifest_dir: PathBuf,
+    working_dir: Option<PathBuf>,
+    manifest_dir: Option<PathBuf>,
     backend: &BackendOptions<()>,
     message_format: MessageFormat,
 ) -> bool {
@@ -326,7 +329,9 @@ fn run_engine(
             ]
             .iter()
             .collect();
-            manifest_dir.join(&relative_path)
+            manifest_dir
+                .map(|manifest_dir| manifest_dir.join(&relative_path))
+                .unwrap_or(relative_path)
         });
 
         let stdout = std::io::BufReader::new(engine_subprocess.stdout.take().unwrap());
@@ -371,10 +376,12 @@ fn run_engine(
                                 .iter()
                                 .map(PathBuf::from)
                                 .map(|path| {
-                                    if path.is_absolute() {
-                                        path
-                                    } else {
+                                    if let Some(working_dir) = working_dir.as_ref()
+                                        && path.is_relative()
+                                    {
                                         working_dir.join(path).to_path_buf()
+                                    } else {
+                                        path
                                     }
                                 })
                                 .map(|path| fs::read_to_string(path).ok())
@@ -640,8 +647,28 @@ fn run_command(options: &Options, haxmeta_files: Vec<EmitHaxMetaMessage>) -> boo
                 path,
             } in haxmeta_files
             {
-                let (haxmeta, id_table): (HaxMeta<Body>, _) =
+                let (mut haxmeta, id_table): (HaxMeta<Body>, _) =
                     HaxMeta::read(fs::File::open(&path).unwrap());
+
+                if let Some(root_module) = &backend.prune_haxmeta {
+                    use hax_frontend_exporter::{DefPathItem, DisambiguatedDefPathItem, IsBody};
+
+                    /// Remove every item from an `HaxMeta` whose path is not `*::<root_module>::**`, where `root_module` is a string.
+                    fn prune_haxmeta<B: IsBody>(haxmeta: &mut HaxMeta<B>, root_module: &str) {
+                        haxmeta.items.retain(|item| match &item.owner_id.path[..] {
+                            [] => true,
+                            [
+                                DisambiguatedDefPathItem {
+                                    data: DefPathItem::TypeNs(s),
+                                    disambiguator: 0,
+                                },
+                                ..,
+                            ] => s == root_module,
+                            _ => false,
+                        })
+                    }
+                    prune_haxmeta(&mut haxmeta, root_module.as_str())
+                }
 
                 error = error
                     || run_engine(
@@ -654,6 +681,12 @@ fn run_command(options: &Options, haxmeta_files: Vec<EmitHaxMetaMessage>) -> boo
                     );
             }
             error
+        }
+        Command::Serialize { .. } => {
+            for EmitHaxMetaMessage { path, .. } in haxmeta_files {
+                HaxMessage::ProducedFile { path, wrote: true }.report(options.message_format, None);
+            }
+            false
         }
     }
 }
@@ -679,7 +712,20 @@ fn main() {
     };
     options.normalize_paths();
 
-    let (haxmeta_files, exit_code) = compute_haxmeta_files(&options);
+    let (haxmeta_files, exit_code) = options
+        .haxmeta
+        .clone()
+        .map(|path| {
+            (
+                vec![EmitHaxMetaMessage {
+                    working_dir: None,
+                    manifest_dir: None,
+                    path,
+                }],
+                0,
+            )
+        })
+        .unwrap_or_else(|| compute_haxmeta_files(&options));
     let error = run_command(&options, haxmeta_files);
 
     std::process::exit(if exit_code == 0 && error {