+++ hax-types/src/cli_options/mod.rs
@@ -341,6 +341,14 @@ pub struct BackendOptions<E: Extension> {
     #[arg(long)]
     pub profile: bool,
 
+    /// Prune Rust items that are not under the provided top-level module name.
+    /// This will effectively remove all items that don't match `*::<prune_haxmetadata>::**`.
+    /// This prunning occurs directly on the `haxmeta` file, in the frontend.
+    /// This is independent from any engine options.
+    #[arg(long)]
+    #[clap(hide = true)]
+    pub prune_haxmeta: Option<String>,
+
     /// Enable engine debugging: dumps the AST at each phase.
     ///
     /// The value of `<DEBUG_ENGINE>` can be either:
@@ -423,6 +431,29 @@ pub enum Command<E: Extension> {
         include_extra: bool,
     },
 
+    /// Serialize to a `haxmeta` file, the internal binary format used by hax to
+    /// store the ASTs produced by the hax exporter.
+    #[clap(hide = true)]
+    Serialize {
+        /// Whether the bodies are exported as THIR, built MIR, const
+        /// MIR, or a combination. Repeat this option to extract a
+        /// combination (e.g. `-k thir -k mir-built`). Pass `--kind`
+        /// alone with no value to disable body extraction.
+        #[arg(
+            value_enum,
+            short,
+            long = "kind",
+            num_args = 0..=3,
+            default_values_t = [ExportBodyKind::Thir]
+        )]
+        kind: Vec<ExportBodyKind>,
+
+        /// When extracting to a given backend, the exporter is called with different `cfg` options.
+        /// This option allows to set the same flags as `cargo hax into` would pick.
+        #[arg(short)]
+        backend: Option<BackendName>,
+    },
+
     #[command(flatten)]
     CliExtension(E::Command),
 }
@@ -431,7 +462,16 @@ impl<E: Extension> Command<E> {
     pub fn body_kinds(&self) -> Vec<ExportBodyKind> {
         match self {
             Command::JSON { kind, .. } => kind.clone(),
-            _ => vec![ExportBodyKind::Thir],
+            Command::Serialize { kind, .. } => kind.clone(),
+            Command::Backend { .. } | Command::CliExtension { .. } => vec![ExportBodyKind::Thir],
+        }
+    }
+    pub fn backend_name(&self) -> Option<BackendName> {
+        match self {
+            Command::Backend(backend_options) => Some((&backend_options.backend).into()),
+            Command::JSON { .. } => None,
+            Command::Serialize { backend, .. } => backend.clone(),
+            Command::CliExtension(_) => None,
         }
     }
 }
@@ -474,6 +514,12 @@ pub struct ExtensibleOptions<E: Extension> {
     /// options like `-C -p <PKG> ;`).
     #[arg(long = "deps")]
     pub deps: bool,
+
+    /// Provide a precomputed haxmeta file explicitly.
+    /// Setting this option bypasses rustc and the exporter altogether.
+    #[arg(long)]
+    #[clap(hide = true)]
+    pub haxmeta: Option<PathBuf>,
 
     /// By default, hax uses `$CARGO_TARGET_DIR/hax` as target folder,
     /// to avoid recompilation when working both with `cargo hax` and
@@ -540,7 +586,7 @@ pub struct ExporterOptions {
 }
 
 #[derive_group(Serializers)]
-#[derive(JsonSchema, Debug, Clone, Copy)]
+#[derive(JsonSchema, ValueEnum, Debug, Clone, Copy)]
 pub enum BackendName {
     Fstar,
     Coq,
@@ -570,22 +616,17 @@ impl fmt::Display for BackendName {
 
 impl From<&Options> for ExporterOptions {
     fn from(options: &Options) -> Self {
-        let backend = match &options.command {
-            Command::Backend(backend_options) => Some((&backend_options.backend).into()),
-            _ => None,
-        };
-        let body_kinds = options.command.body_kinds();
         ExporterOptions {
             deps: options.deps,
             force_cargo_build: options.force_cargo_build.clone(),
-            backend,
-            body_kinds,
+            backend: options.command.backend_name(),
+            body_kinds: options.command.body_kinds(),
         }
     }
 }
 
-impl From<&Backend<()>> for BackendName {
-    fn from(backend: &Backend<()>) -> Self {
+impl<E: Extension> From<&Backend<E>> for BackendName {
+    fn from(backend: &Backend<E>) -> Self {
         match backend {
             Backend::Fstar { .. } => BackendName::Fstar,
             Backend::Coq { .. } => BackendName::Coq,